Let user input a data lag when deploying a model
