+++ src/components/modelDeployment/deploymentUtils.jsx
@@ -176,3 +176,54 @@ async function getStateMachineStatus(gateway, arnList) {
 
     return schedulerLaunched
 }
+
+// -----------------------------------------------------
+// Get the sampling rate of the model passed as argument
+// -----------------------------------------------------
+export async function getModelSamplingRate(gateway, modelName) {
+    if (!modelName) { return undefined }
+    const response = await gateway.lookoutEquipment
+                            .describeModel(modelName)
+                            .catch((error) => { console.log(error.response) })
+
+    const possibleSamplingRate = {
+        'PT1S': 1, 
+        'PT5S': 5,
+        'PT10S': 10,
+        'PT15S': 15,
+        'PT30S': 30,
+        'PT1M': 60,
+        'PT5M': 300,
+        'PT10M': 600,
+        'PT15M': 900,
+        'PT30M': 1800,
+        'PT1H': 3600
+    }
+    let modelSamplingRate = possibleSamplingRate[response['DataPreProcessingConfiguration']['TargetSamplingRate']]
+
+    const possibleSchedulerSamplingRate = {
+        300: { label: "5 minutes", value: "PT5M" },
+        600: { label: "10 minutes", value: "PT10M" },
+        900: { label: "15 minutes", value: "PT15M" },
+        1800: { label: "30 minutes", value: "PT30M" },
+        3600: { label: "60 minutes", value: "PT1H" }
+    }
+
+    let schedulerSROptions = []
+    let selectedOption = undefined
+    const srKeys = Object.keys(possibleSchedulerSamplingRate)
+    srKeys.forEach((sr) => {
+        if (sr >= modelSamplingRate) {
+            schedulerSROptions.push(possibleSchedulerSamplingRate[sr])
+            if (!selectedOption) {
+                selectedOption = possibleSchedulerSamplingRate[sr]
+            }
+        }
+    })
+
+    return {
+        sr: modelSamplingRate,
+        srOptions: schedulerSROptions,
+        selectedOption: selectedOption
+    }
+}
